+++ src/Component/forAdmin/Admin.css
@@ -1,0 +1,164 @@
+:root {
+  --main-color: #0652dd;
+  --text-color: #455560;
+  --secondary-color: #cc2e5d;
+}
+
+.page-wrap {
+  display: flex;
+  min-height: 100vh;
+}
+
+.left-panel {
+  flex: 1 1 66.666%;
+  background: url("https://previews.123rf.com/images/loonara/loonara1504/loonara150400070/38213792-hipster-home-office-tabletop-papers-and-notebooks-reading-glasses-smart-phone-pen-on-an-old-wooden.jpg")
+    no-repeat;
+  display: flex;
+  /* flex-direction: column; */
+  justify-content: center;
+  align-items: center;
+}
+
+.left-panel .illustration {
+  margin-bottom: 50px;
+}
+
+.left-panel h2 {
+  color: #fff;
+  font-size: 48px;
+  font-weight: 400;
+  margin-bottom: 0px;
+  border-radius: 5px;
+  margin-top: 153px;
+
+  text-shadow: 0 0 0.2em rgb(5, 56, 242);
+}
+
+.right-panel {
+  flex: 1 1 33.333%;
+  padding: 25px;
+}
+
+.right-panel h1 {
+  font-family: "Times New Roman", Times, serif;
+  /* color: var(--main-color); */
+  font-size: 48px;
+  font-weight: 400;
+  margin-bottom: 50px;
+
+  color: transparent;
+  background-clip: text;
+  -webkit-background-clip: text;
+  background-image: linear-gradient(
+    to right,
+    rgb(193, 62, 216),
+    rgb(73, 22, 167)
+  );
+}
+
+.right-panel h1 strong {
+  color: transparent;
+  background-clip: text;
+  -webkit-background-clip: text;
+  background-image: linear-gradient(
+    to right,
+    rgb(196, 58, 220),
+    rgb(100, 21, 246)
+  );
+  font-weight: 600;
+}
+
+.animated-form h3 {
+  color: var(--text-color);
+  font-size: 28px;
+  font-weight: 500;
+  margin-bottom: 45px;
+}
+
+.animated-form .form-group {
+  position: relative;
+  display: flex;
+  margin-bottom: 45px;
+}
+
+.form-group.flex-end {
+  justify-content: flex-end;
+}
+
+.animated-form .form-group label {
+  position: absolute;
+  top: 50%;
+  left: 0px;
+  transform: translateY(-50%);
+  color: #aaa;
+  font-size: 20px;
+  transition: 0.4s ease-out;
+}
+
+.animated-form .form-group input {
+  display: block;
+  width: 100%;
+  padding: 10px 0px;
+  border: none;
+  outline: none;
+  background: none;
+  border-bottom: 3px solid #aaa;
+  color: var(--text-color);
+  font-size: 20px;
+  transition: 0.4s ease-out;
+}
+
+.animated-form .form-group:focus-within label,
+.animated-form .form-group.has-content label {
+  top: 0px;
+  transform: translateY(-100%);
+  color: var(--main-color);
+}
+
+.animated-form .form-group:focus-within input,
+.animated-form .form-group.has-content input {
+  border-bottom-color: var(--main-color);
+}
+
+.animated-form .form-group.flex-end {
+  margin-bottom: 25px;
+}
+
+.animated-form .form-group .button {
+  display: inline-block;
+  width: auto;
+  padding: 15px 60px;
+  border: 3px solid var(--main-color);
+  border-radius: 999px;
+  background-image: linear-gradient(
+    to right,
+    transparent 50%,
+    var(--secondary-color) 50%,
+    var(--main-color)
+  );
+  background-size: 200%;
+  color: var(--main-color);
+  font-size: 24px;
+  font-weight: 500;
+  cursor: pointer;
+  text-decoration: none;
+}
+
+.animated-form .form-group .button:hover {
+  color: #fff;
+  background-position: 100%;
+  border: 3px solid #fff;
+}
+
+.animated-form .lost-password a {
+  color: var(--main-color);
+  font-size: 16px;
+  font-weight: 400;
+  text-decoration: none;
+  padding-bottom: 5px;
+  border-bottom: 2px solid transparent;
+}
+
+.animated-form .lost-password a:hover {
+  border-bottom-color: var(--main-color);
+}
